+++ compiler-plugin-tests/src/test/java/io/ballerina/stdlib/sql/compiler/CompilerPluginTest.java
@@ -36,8 +36,6 @@ import java.nio.file.Paths;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static io.ballerina.stdlib.sql.compiler.SQLDiagnosticsCodes.SQL_901;
-
 /**
  * Tests the custom SQL compiler plugin.
  */
@@ -119,32 +117,6 @@ public class CompilerPluginTest {
         DiagnosticInfo timeWithTimezoneOutParameter = errorDiagnosticsList.get(3).diagnosticInfo();
         Assert.assertEquals(timeWithTimezoneOutParameter.code(), SQLDiagnosticsCodes.SQL_231.getCode());
         Assert.assertEquals(timeWithTimezoneOutParameter.messageFormat(), SQLDiagnosticsCodes.SQL_231.getMessage());
-    }
-
-    @Test
-    public void testOutParameterHint() {
-        Package currentPackage = loadPackage("sample3");
-        PackageCompilation compilation = currentPackage.getCompilation();
-        DiagnosticResult diagnosticResult = compilation.diagnosticResult();
-        List<Diagnostic> errorDiagnosticsList = diagnosticResult.diagnostics().stream()
-                .filter(r -> r.diagnosticInfo().severity().equals(DiagnosticSeverity.ERROR))
-                .collect(Collectors.toList());
-        long availableErrors = errorDiagnosticsList.size();
-
-        Assert.assertEquals(availableErrors, 2);
-
-        List<Diagnostic> hintDiagnosticsList = diagnosticResult.diagnostics().stream()
-                .filter(r -> r.diagnosticInfo().severity().equals(DiagnosticSeverity.HINT))
-                .collect(Collectors.toList());
-        long availableHints = hintDiagnosticsList.size();
-
-        Assert.assertEquals(availableHints, 2);
-
-        hintDiagnosticsList.forEach(diagnostic -> {
-            Assert.assertEquals(diagnostic.diagnosticInfo().code(), SQL_901.getCode());
-            Assert.assertEquals(diagnostic.diagnosticInfo().messageFormat(), SQL_901.getMessage());
-        });
-
     }
 
     @Test